fix: vocab_size (used by model) should include added tokens
This ones a doozy, HF defines vocab size as not including the added
tokens, and the length of the tokenizer as including the added tokens
But the model's "vocab_size" needs include the added tokens (cause it
needs to do predictions over *all* tokens)

This hasn't been an issue yet, cause smirk and MolFormer include all
tokens in their vocab, but smirk-gpe separates them.

This commit adds test to confirm that len is being used to set
vocab_size on the datamodule using len and the expected relationship
between vocab_size, len with/without added tokens. Plus fixes the issue
for both dataloaders