update priors
Estimate lr_0 as ~1e-3 and lr_n as ~0.1; this gives ~1.6e-4 for ~29M
parameters (Similar to MolFormer). Plus change the prior for `lr_n` to
Exponential, to reflect belief that a smaller scaling is more likely that
a larger one

Update priors on ff/aspect shape factor base on results to-date (shape
factors have limited impact on model performance).

Also clean up the code a bit (formatting)